Fixed Commands for Auto :smiling_imp: :rainbow:
