+++ src/main/java/frc/robot/Robot.java
@@ -65,6 +65,8 @@ public class Robot extends TimedRobot {
 
     if (m_autonomousCommand != null) {
       m_autonomousCommand.schedule();
+
+
     }
 
   }
